Add additional processing for migrating queries to titles
First we try to determine the title for the search query:
1. If the query contains both a search string and arguments then
use the search string as title.
2. If the query is for an author (when using follow author) then use
the author name as the title.
3. Else fall back to the search query as a title.

Then we try to convert the saved search to an actual search query.
Facets are converted to search strings and combined with the initial
query. Sorts are discarded as that cannot be handled by 
Follow searches at the moment.